+++ packages/remote-core/src/connection/connectHostRenderRoot.ts
@@ -12,6 +12,7 @@ import { ThreadNestedIframe } from "@quilted/threads";
 interface Options {
   root: HTMLDivElement;
   onPathnameChanged?: (pathname: string) => void;
+  packageVersion?: string;
 }
 
 const incompatibleParentFrameError = () =>
@@ -31,7 +32,7 @@ export const connectRemoteReceiver = (
 export const connectHostRenderRoot = async (
   options: Options,
 ): Promise<RemoteToHostConnection> => {
-  const { root, onPathnameChanged } = options;
+  const { root, onPathnameChanged, packageVersion } = options;
 
   const connection = new ThreadNestedIframe<HostExports, RemoteExports>({
     serialization: new FlowThreadSerialization(),
@@ -49,7 +50,10 @@ export const connectHostRenderRoot = async (
   }
 
   try {
-    await connection.imports.setIsReady(Version.v3);
+    await connection.imports.setIsReady({
+      version: Version.v4,
+      packageVersion,
+    });
 
     if (typeof mwExtBridge !== "undefined") {
       mwExtBridge.connection = connection.imports;